requirements.txt
chainlit
speechrecognition
google-generativeai
pytube
python-dotenv
-e .